piscine101/python03/ex01/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
discord.py==1.7.3
Django==3.2.4
idna==3.2
image==1.5.33
multidict==5.1.0
Pillow==8.2.0
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
yarl==1.6.3
